backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
ultralytics
pillow
numpy
opencv-python
requests
flask
werkzeug
